Autenticacion admin/deletebook(Agregamos la autenticacion del adminis…

+++ src/admin/DeleteBookAdmin.ts
@@ -10,9 +10,13 @@ export const deleteBookAdmin = async (
     const { id } = req.params;
 
     try {
+        const book = await Libros.findByPk(id);
+        if(!book){
+            return res.status(404).send('Libro no encontrado')
+        }
         await Libros.destroy({ where: { id } });
         await LibrosGeneros.destroy({ where: { libroid: id } });
-        res.json('Libro eliminado');
+        return res.json('Libro eliminado');
     } catch (err) {
         next(err);
     }
